test: move integration _utils.py to shared tests/_utils.py
