+++ tests/integration/test_actor.py
@@ -6,7 +6,7 @@ from collections.abc import AsyncIterator, Iterator
 from datetime import UTC, datetime, timedelta
 from typing import TYPE_CHECKING
 
-from ._utils import get_random_resource_name, maybe_await
+from .._utils import get_random_resource_name, maybe_await
 from apify_client._models import (
     Actor,
     ActorChargeEvent,
