+++ tests/integration/test_actor_env_var.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from collections.abc import AsyncIterator, Iterator
 from typing import TYPE_CHECKING
 
-from ._utils import get_random_resource_name, maybe_await
+from .._utils import get_random_resource_name, maybe_await
 from apify_client._models import Actor, EnvVar, ListOfEnvVars
 
 if TYPE_CHECKING:
