+++ tests/integration/test_key_value_store.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING
 import impit
 import pytest
 
-from ._utils import (
+from .._utils import (
     KvsFixture,
     collect_iterate_until_present,
     get_random_resource_name,
